Fixed GenAI Exporting & Importing Mechanism
