+++ lib/models/settings_model.dart
@@ -1,5 +1,5 @@
 import 'package:apidash_core/apidash_core.dart';
-import 'package:genai/llm_saveobject.dart';
+import 'package:genai/genai.dart';
 import 'package:flutter/material.dart';
 import 'package:apidash/consts.dart';
