+++ lib/providers/collection_providers.dart
@@ -1,12 +1,9 @@
 import 'dart:convert';
 import 'package:apidash_core/apidash_core.dart';
 import 'package:genai/genai.dart';
-import 'package:genai/models/ai_request_model.dart';
-import 'package:genai/models/ai_response_model.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:apidash/consts.dart';
-import 'package:genai/providers/common.dart';
 import 'providers.dart';
 import '../models/models.dart';
 import '../services/services.dart' show hiveHandler, HiveHandler;
