+++ lib/screens/history/history_widgets/ai_history_page.dart
@@ -1,12 +1,10 @@
-import 'package:apidash/providers/collection_providers.dart';
 import 'package:apidash/providers/history_providers.dart';
+import 'package:apidash/screens/home_page/editor_pane/details_card/request_pane/ai_request/widgets/ai_config_widgets.dart';
 import 'package:apidash/widgets/editor.dart';
 import 'package:apidash_design_system/tokens/measurements.dart';
-import 'package:apidash_design_system/widgets/textfield_outlined.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:genai/llm_config.dart';
-import 'package:genai/widgets/ai_config_widgets.dart';
+import 'package:genai/genai.dart';
 
 class HisAIRequestPromptSection extends ConsumerWidget {
   const HisAIRequestPromptSection({super.key});
