+++ lib/screens/home_page/editor_pane/details_card/request_pane/ai_request/widgets/ai_config_widgets.dart
@@ -1,6 +1,6 @@
 import 'package:apidash_design_system/widgets/textfield_outlined.dart';
 import 'package:flutter/material.dart';
-import 'package:genai/llm_config.dart';
+import 'package:genai/genai.dart';
 
 class SliderAIConfig extends StatelessWidget {
   final LLMModelConfiguration configuration;
@@ -35,7 +35,8 @@ class SliderAIConfig extends StatelessWidget {
           ),
         ),
         Text(
-          (configuration.configValue.value as (double, double, double)).$2
+          (configuration.configValue.value as (double, double, double))
+              .$2
               .toStringAsFixed(2),
         ),
       ],
