+++ lib/screens/home_page/editor_pane/url_card.dart
@@ -1,11 +1,7 @@
+import 'package:apidash/screens/home_page/editor_pane/details_card/request_pane/ai_request/widgets/llm_selector.dart';
 import 'package:apidash_core/apidash_core.dart';
-
+import 'package:genai/genai.dart';
 import 'package:apidash_design_system/apidash_design_system.dart';
-import 'package:genai/llm_input_payload.dart';
-import 'package:genai/llm_saveobject.dart';
-import 'package:genai/models/ai_request_model.dart';
-
-import 'package:genai/widgets/llm_selector.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:apidash/providers/providers.dart';
