+++ lib/screens/settings_page.dart
@@ -1,4 +1,4 @@
-import 'package:genai/widgets/llm_selector.dart';
+import 'package:apidash/screens/home_page/editor_pane/details_card/request_pane/ai_request/widgets/llm_selector.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
