+++ packages/genai/lib/providers/azureopenai/models.dart
@@ -1,9 +1,6 @@
 // ignore_for_file: constant_identifier_names
-
-import 'package:genai/llm_config.dart';
-import 'package:genai/providers/common.dart';
-import 'package:genai/providers/gemini/gemini.dart';
-import 'package:genai/providers/providers.dart';
+import '../../providers/common.dart';
+import '../providers.dart';
 
 enum AzureOpenAIModel implements LLMModel {
   custom('custom', 'Custom', LLMProvider.azureopenai);
